Revert "saving authenticated user data in redis"
This reverts commit 66ce5341
