+++ src/main/java/com/dnetwork/DnetworkApplication.java
@@ -1,10 +1,38 @@
 package com.dnetwork;
 
-
+import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-
+import org.springframework.context.annotation.Bean;
+import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.serializer.GenericToStringSerializer;
 
 @SpringBootApplication
 public class DnetworkApplication {
+
+    @Bean
+    JedisConnectionFactory jedisConnectionFactory(){
+        return new JedisConnectionFactory();
+    }
+
+//    @Bean
+//    RedisTemplate<String, Object> redisTemplate() {
+//        RedisTemplate<String, Object> redisTemplate = new RedisTemplate<>();
+//        redisTemplate.setConnectionFactory(jedisConnectionFactory());
+//        return redisTemplate;
+//    }
+
+
+    @Bean
+    public RedisTemplate<String, Object> redisTemplate() {
+        final RedisTemplate<String, Object> template = new RedisTemplate<String, Object>();
+        template.setConnectionFactory(jedisConnectionFactory());
+        template.setValueSerializer(new GenericToStringSerializer<>(Object.class));
+        return template;
+    }
+
+    public static void main(String[] args) {
+        SpringApplication.run(DnetworkApplication.class, args);
+    }
 
 }
