--- src/main/java/com/dnetwork/config/RedisConfig.java
@@ -1,31 +1,0 @@
-package com.dnetwork.config;
-
-import com.dnetwork.DnetworkApplication;
-import org.springframework.boot.SpringApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.serializer.GenericToStringSerializer;
-
-@Configuration
-public class RedisConfig {
-
-    @Bean
-    JedisConnectionFactory jedisConnectionFactory(){
-        return new JedisConnectionFactory();
-    }
-
-    @Bean
-    public RedisTemplate<String, Object> redisTemplate() {
-        final RedisTemplate<String, Object> template = new RedisTemplate<String, Object>();
-        template.setConnectionFactory(jedisConnectionFactory());
-        template.setValueSerializer(new GenericToStringSerializer<>(Object.class));
-        return template;
-    }
-
-    public static void main(String[] args) {
-        SpringApplication.run(DnetworkApplication.class, args);
-    }
-
-}
